services/outbox_poller/requirements.txt
Flask
psycopg2-binary
confluent-kafka